% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{sim_tables}
\alias{sim_tables}
\title{Simulate fake data with a fixed difference between groups}
\usage{
sim_tables(alphaA = 0.5, betaA = 5, delta = 0.1, alphaB = NULL,
  betaB = NULL, N = 10, nA = 10000, nB = 100)
}
\arguments{
\item{alphaA}{true parameter of prior Beta-distribution for group A}

\item{betaA}{true parameter of prior Beta-distribution for group A}

\item{delta}{true difference in probability between groups}

\item{alphaB}{true parameter of prior Beta-distribution for group B}

\item{betaB}{true parameter of prior Beta-distribution for group B}

\item{N}{number of contingency tables}

\item{nA}{number of tries group A}

\item{nB}{number of tries group B}
}
\value{
Data frame of fake data and true probabilities
}
\description{
The function can either simulate the groups from one beta distribution
and setting afixed difference (delta) between them, or by drawing from
two different Beta distributions. Set delta to NULL and give parameter
values for (alphaB, betaB) to activate the second behaviour.
}
