% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{tables_to_df}
\alias{tables_to_df}
\title{Turn list of 2 x 2 contingency tables into data frame.}
\usage{
tables_to_df(ctables)
}
\arguments{
\item{ctables}{a list of matrices or data frames}
}
\value{
A data frame where each row corresponds to a contingency table.
}
\description{
This convenience function turns a list of 2 x 2 matrices into a data frame.
There should be two columns each containing a group. The first row should
contain the number of successes, and the second row the number of tries.
}
